requirements.txt
crispy-bootstrap5==2026.3
django==5.2.7
django-crispy-forms==2.6
django-filter==25.2
openpyxl==3.1.5
psycopg[binary]==3.3.3
python-dotenv==1.2.2
